🐛 fix: fix problem of can't update Post Like when login or logout
