+++ src/components/Recommend/FrameOrganism/FrameOrganism.tsx
@@ -1,6 +1,6 @@
 import {useNavigation} from '@react-navigation/native';
-import React from 'react';
-import {useDispatch} from 'react-redux';
+import React, {useEffect} from 'react';
+import {useDispatch, useSelector} from 'react-redux';
 
 import {ButtonText, TextnIconWrapper} from '../PoseOrganism/PoseOrganism.styles';
 import RecommendPreviewFourCard from '../PreviewSixCard';
@@ -16,15 +16,26 @@ import {
 
 import useGetInfinitePosts from 'src/querys/useGetInfinitePosts';
 import {changeFilteredFrame} from 'src/redux/actions/PostAction';
+import {RootState} from 'src/redux/store';
 
 const FrameRecommendOrganism = () => {
   const dispatch = useDispatch();
-  const {data, isLoading} = useGetInfinitePosts({tagIdSet: [41], order: 'popular', key: 'frame'});
+  const {accessToken} = useSelector((state: RootState) => state.userReducer);
+  const {data, isLoading, refetch} = useGetInfinitePosts({
+    tagIdSet: [41],
+    order: 'popular',
+    key: 'frame',
+  });
   const navigation = useNavigation();
 
   const handlePressCard = (id: number) => () => {
     navigation.navigate('RecommendDetail' as never, {postId: id} as never);
   };
+
+  useEffect(() => {
+    refetch();
+  }, [accessToken]);
+
   return (
     <OrganismView>
       <TitleWrapper>
