+++ src/components/UserProvider/index.ts
@@ -1,0 +1,3 @@
+import UserProvider from './UserProvider';
+
+export default UserProvider;
