+++ src/components/utils/Interceptor/index.tsx
@@ -5,7 +5,7 @@ import EncryptedStorage from 'react-native-encrypted-storage';
 import {useDispatch, useSelector} from 'react-redux';
 
 import getAccessToken from 'src/apis/getAccessToken';
-import {loginAction, setAccessToken, setInterceptor} from 'src/redux/actions/UserAction';
+import {loginAction, setAccessToken} from 'src/redux/actions/UserAction';
 import {RootState} from 'src/redux/store';
 import getApiServer from 'src/utils/getApiServer';
 
@@ -17,60 +17,55 @@ const Interceptor = ({children}: PropsWithChildren<{}>) => {
   const dispatch = useDispatch();
   const {accessToken} = useSelector((state: RootState) => state.userReducer);
 
-  useEffect(() => {
-    if (!accessToken) {
-      return;
-    }
-
-    const requestInterceptor = AxiosInstance.interceptors.request.use(
-      config => {
-        if (!config.headers) {
-          return config;
-        }
-
-        if (accessToken) {
-          config.headers.Authorization = `Bearer ${accessToken}`;
-        }
+  const requestInterceptor = AxiosInstance.interceptors.request.use(
+    config => {
+      if (!config.headers) {
         return config;
-      },
-      error => {
-        return Promise.reject(error);
-      },
-    );
-    const responseInterceptor = AxiosInstance.interceptors.response.use(
-      response => {
-        return response;
-      },
-      async error => {
-        const {
-          config,
-          response: {status, data},
-        } = error;
-        if (status === 403) {
-          const originalHeader = config;
-          const refreshToken = await EncryptedStorage.getItem('refreshToken');
-          if (!refreshToken) {
-            return;
-          }
-          if (data.code === -100012) {
-            EncryptedStorage.removeItem('refreshToken');
-            Alert.alert('로그인이 만료되어 재로그인이 필요합니다.');
-          }
-          const newAccessToken = await getAccessToken(refreshToken);
-          originalHeader.headers.Authorization = `Bearer ${newAccessToken}`;
-          dispatch(setAccessToken(newAccessToken));
-          dispatch(loginAction(true));
-          return axios(originalHeader);
+      }
+      if (accessToken) {
+        config.headers.Authorization = `Bearer ${accessToken}`;
+      }
+      return config;
+    },
+    error => {
+      return Promise.reject(error);
+    },
+  );
+  const responseInterceptor = AxiosInstance.interceptors.response.use(
+    response => {
+      return response;
+    },
+    async error => {
+      const {
+        config,
+        response: {status, data},
+      } = error;
+      if (status === 403) {
+        const originalHeader = config;
+        const refreshToken = await EncryptedStorage.getItem('refreshToken');
+        if (!refreshToken) {
+          return;
         }
-        return Promise.reject(error);
-      },
-    );
-    dispatch(setInterceptor());
+        if (data.code === -100012) {
+          EncryptedStorage.removeItem('refreshToken');
+          Alert.alert('로그인이 만료되어 재로그인이 필요합니다.');
+        }
+        const newAccessToken = await getAccessToken(refreshToken);
+        originalHeader.headers.Authorization = `Bearer ${newAccessToken}`;
+        dispatch(setAccessToken(newAccessToken));
+        dispatch(loginAction(true));
+        return axios(originalHeader);
+      }
+      return Promise.reject(error);
+    },
+  );
+
+  useEffect(() => {
     return () => {
       AxiosInstance.interceptors.request.eject(requestInterceptor);
       AxiosInstance.interceptors.response.eject(responseInterceptor);
     };
-  }, [accessToken]);
+  });
 
   return <>{children}</>;
 };
