+++ src/hooks/useLogout.ts
@@ -1,7 +1,6 @@
 import {GoogleSignin} from '@react-native-google-signin/google-signin';
 import {logout} from '@react-native-seoul/kakao-login';
-import EncryptedStorage from 'react-native-encrypted-storage/';
-import {useQueryClient} from 'react-query';
+import EncryptedStorage from 'react-native-encrypted-storage';
 import {useDispatch, useSelector} from 'react-redux';
 
 import {changeUserInfo, loginAction, setAccessToken} from 'src/redux/actions/UserAction';
@@ -9,7 +8,6 @@ import {RootState} from 'src/redux/store';
 
 const useLogout = () => {
   const dispatch = useDispatch();
-  const queryClient = useQueryClient();
   const {userInfo} = useSelector((state: RootState) => state.userReducer);
 
   const signOut = async () => {
@@ -24,7 +22,6 @@ const useLogout = () => {
     dispatch(setAccessToken(''));
     dispatch(changeUserInfo({}));
     dispatch(loginAction(false));
-    queryClient.invalidateQueries();
   };
 
   return signOut;
