+++ src/redux/reducers/userReducer.ts
@@ -1,10 +1,9 @@
 import {Reducer} from 'redux';
 
-import {LOGIN, USER_INFO, SET_ACCESS_TOKEN, SET_INTERCEPTOR} from '../types/UserActionType';
+import {LOGIN, USER_INFO, SET_ACCESS_TOKEN} from '../types/UserActionType';
 
 const initialState = {
   isLoggedIn: false,
-  isSettingInterceptor: false,
   accessToken: '',
   userInfo: {},
 };
@@ -19,8 +18,6 @@ const tabBarReducer: Reducer = (state = initialState, action) => {
       return {...state, userInfo: {...payload.userInfo}};
     case SET_ACCESS_TOKEN:
       return {...state, accessToken: payload.accessToken};
-    case SET_INTERCEPTOR:
-      return {...state, isSettingInterceptor: true};
     default:
       return state;
   }
