+++ src/redux/types/UserActionType.ts
@@ -1,4 +1,3 @@
 export const LOGIN = 'LOGIN';
 export const USER_INFO = 'USER_INFO';
 export const SET_ACCESS_TOKEN = 'SET_ACCESS_TOKEN';
-export const SET_INTERCEPTOR = 'SET_INTERCEPTOR';
